feat: :sparkles: add userController to app to divide code up more
This will make the routes files more readable and consolidates the controller logic separate from routes
